Remove `irnetbox-proxy` tool
We want to focus this project on the stbt *library*. The surrounding
tooling is becoming a maintenance burden, and for all I know it doesn't
work properly because it's hard to write tests for it.

This particular tool serves to work around a limitation of the RedRat
irNetBox (mk 3, at least) which doesn't handle multiple simultaneous
clients gracefully. If you want to use it feel free to maintain it in a
separate repo as per the [license].

[license]: https://github.com/stb-tester/stb-tester/blob/master/LICENSE