tests: Fix hang (due to gst source going away) when test fails
We run `test.py` writing video to a pipe, and `verify.py` reading video
from that pipe. If the test fails, the `test.py` process finishes first
so the `verify.py` process says:

> verify.py: warning: Got EOS from source pipeline
> verify.py: warning: Attempting to recover from video loss: Stopping source pipeline and waiting 5s...
> verify.py: warning: Restarting source pipeline...

...and then it hangs.

[Note that this behaviour always happens, regardless of the config
setting `global.restart_source` which controls restarting the source
when we get underruns without receiving an explicit EOS.]

When the test passes, `verify.py` finished much earlier that `test.py`
so we wouldn't see this happen.

I'm not too concerned that this is a real bug because it's only seen
with --source-pipeline='filesrc location=fifo' which is an artificial
setup only used in our self-tests.