Fix test_template_annotation_with_ndarray_template
On my laptop it has started matching at the top of the videotestsrc
frame, so the annotation with the name of the template isn't visible.

Now the test is identical to `test_template_annotation_labels` except
that we load the reference image with `cv2.imread` so that it is a raw
numpy array.