stbt.match video annotation: Draw "match(<Custom Image>)" instead of...
...instead of "match(u'<Custom Image>')".

This makes it more repr-like (because <Custom Image> isn't a string,
it's a placeholder for an object). I needed to change
`custom-image-label.png` anyway because the test was expecting
`'<Custom Image>'` (with single quotes) but it was getting
`u'<Custom Image>'` (with u + single quotes). Now it expects
`<Custom Image>` (without any kind of quote or backticks).